#3741 Fixed an issue where time filters were not applied
